requirements.txt
# let me know if i've missed anything
gradio
jsonschema
matplotlib
numba
numpy
opencv-python
Pillow
scipy
timm==0.6.13
torch
torchvision
tqdm
